+++ docs/pages/documentation/form/_Form.vue
@@ -1,6 +1,7 @@
 <template>
     <div class="nav-left">
         <router-link to="input" class="nav-item is-tab">Input</router-link>
+        <router-link to="select" class="nav-item is-tab">Select</router-link>
         <router-link to="checkbox" class="nav-item is-tab">Checkbox</router-link>
         <router-link to="radio" class="nav-item is-tab">Radio</router-link>
         <router-link to="switch" class="nav-item is-tab">Switch</router-link>