+++ docs/router/index.js
@@ -25,6 +25,7 @@ const Table = (r) => require(['../pages/documentation/data/Table'], r)
 const Pagination = (r) => require(['../pages/documentation/data/Pagination'], r)
 
 const Input = (r) => require(['../pages/documentation/form/Input'], r)
+const Select = (r) => require(['../pages/documentation/form/Select'], r)
 const Checkbox = (r) => require(['../pages/documentation/form/Checkbox'], r)
 const Radio = (r) => require(['../pages/documentation/form/Radio'], r)
 const Switch = (r) => require(['../pages/documentation/form/Switch'], r)
@@ -136,6 +137,13 @@ export default new Router({
                 {
                     path: 'input',
                     component: Input,
+                    meta: {
+                        category: 'FormElements'
+                    }
+                },
+                {
+                    path: 'select',
+                    component: Select,
                     meta: {
                         category: 'FormElements'
                     }